+++ Sources/MapboxNavigationCore/Navigator/Internals/NativeHandlersFactory.swift
@@ -221,6 +221,7 @@ final class NativeHandlersFactory: @unchecked Sendable {
         let defaultConfig = [
             customConfigFeaturesKey: [
                 "useInternalReroute": true,
+                "useInternalRouteRefresh": true,
                 "useTelemetryNavigationEvents": true,
             ],
             "navigation": [
@@ -233,10 +234,10 @@ final class NativeHandlersFactory: @unchecked Sendable {
         ]
 
         var customConfig = UserDefaults.standard.dictionary(forKey: customConfigKey) ?? [:]
-        customConfig.deepMerge(with: defaultConfig, uniquingKeysWith: { first, _ in first })
+        customConfig.deepMerge(with: defaultConfig, uniquingKeysWith: { _, defaultConfigValue in defaultConfigValue })
 
         let customConfigJSON: String
-        if let jsonDataConfig = try? JSONSerialization.data(withJSONObject: customConfig, options: []),
+        if let jsonDataConfig = try? JSONSerialization.data(withJSONObject: customConfig, options: [.sortedKeys]),
            let encodedConfig = String(data: jsonDataConfig, encoding: .utf8)
         {
             customConfigJSON = encodedConfig